7.pizza_redis_pipeline/second-attempt/dashboard/requirements.txt
pymongo
redis
streamlit
plotly
pandas
python-dotenv
